Home page and NASA project page
The home page and NASA project page have been completed. Temporary/Placeholder for other pages have also been created for other areas of the website. The framework of the website has been created and a template html page has been created to copy for new pages.

The original plan for the website was to load all of the sections in <div>'s and choose which div to show based on the item clicked. This idea was scrapped because of the high amount of system resources required to keep many pages worth of data in memory. Instead, sections will be separated properly into different pages to reduce total resources needed at time of loading for each page.

The CitySmart video demonstration is too large at the moment to be added using GitHub desktop and will be added at a later.