Addition of the Monopoly Page and Bug Fixes
I created the Monopoly page and added photos to the project that could be used on the page. This is only a rough draft and the page will be edited further down the road. I also fixed a few pages that underlined the entire div of the word "Home" instead of just the word. Other various small changes as well.